Move scm option types and corresponding valid() checks into a single file
