+++ scm/helper.go
@@ -78,54 +78,6 @@ var (
 	ErrAlreadyExists = errors.New("course repositories already exist for that organization: " + repoNames)
 )
 
-// Validators //
-
-func (opt GetOrgOptions) valid() bool {
-	return opt.ID != 0 || opt.Name != ""
-}
-
-func (opt UpdateEnrollmentOptions) valid() bool {
-	return opt.Organization != "" && opt.User != ""
-}
-
-func (opt *RejectEnrollmentOptions) valid() bool {
-	return opt.OrganizationID > 0 && opt.RepositoryID > 0 &&
-		opt.User != ""
-}
-
-func (opt UpdateTeamOptions) valid() bool {
-	return opt.TeamID > 0 && opt.OrganizationID > 0
-}
-
-func (opt CreateRepositoryOptions) valid() bool {
-	return opt.Organization != "" && opt.Path != ""
-}
-
-func (opt TeamOptions) valid() bool {
-	return opt.TeamName != "" && opt.Organization != ""
-}
-
-func (opt RepositoryOptions) valid() bool {
-	return opt.ID > 0 || (opt.Path != "" && opt.Owner != "")
-}
-
-func (opt *IssueOptions) valid() bool {
-	return opt.Organization != "" && opt.Repository != "" && opt.Title != "" && opt.Body != ""
-}
-
-func (opt RequestReviewersOptions) valid() bool {
-	return opt.Organization != "" && opt.Repository != "" &&
-		opt.Number > 0 && len(opt.Reviewers) != 0
-}
-
-func (opt IssueCommentOptions) valid() bool {
-	return opt.Organization != "" && opt.Repository != "" && opt.Body != ""
-}
-
-func (opt InvitationOptions) valid() bool {
-	return opt.Login != "" && opt.Owner != "" && opt.Token != ""
-}
-
 // Errors //
 
 // ErrNotSupported is returned when the source code management solution used
