+++ scm/mock.go
@@ -43,7 +43,7 @@ func NewMockSCMClient() *MockSCM {
 
 // Clone copies the repository in testdata to the given destination path.
 func (s MockSCM) Clone(ctx context.Context, opt *CloneOptions) (string, error) {
-	if _, err := s.GetOrganization(ctx, &GetOrgOptions{
+	if _, err := s.GetOrganization(ctx, &OrganizationOptions{
 		Name: opt.Organization,
 	}); err != nil {
 		return "", err
@@ -58,7 +58,7 @@ func (s MockSCM) Clone(ctx context.Context, opt *CloneOptions) (string, error) {
 }
 
 // GetOrganization implements the SCM interface.
-func (s *MockSCM) GetOrganization(ctx context.Context, opt *GetOrgOptions) (*qf.Organization, error) {
+func (s *MockSCM) GetOrganization(ctx context.Context, opt *OrganizationOptions) (*qf.Organization, error) {
 	if !opt.valid() {
 		return nil, fmt.Errorf("invalid argument: %+v", opt)
 	}
@@ -90,7 +90,7 @@ func (s *MockSCM) CreateRepository(ctx context.Context, opt *CreateRepositoryOpt
 	if !opt.valid() {
 		return nil, fmt.Errorf("invalid argument: %+v", opt)
 	}
-	org, err := s.GetOrganization(ctx, &GetOrgOptions{
+	org, err := s.GetOrganization(ctx, &OrganizationOptions{
 		Name: opt.Organization,
 	})
 	if err != nil {
@@ -157,7 +157,7 @@ func (s *MockSCM) CreateIssue(ctx context.Context, opt *IssueOptions) (*Issue, e
 	if !opt.valid() {
 		return nil, fmt.Errorf("invalid argument: %v", opt)
 	}
-	if _, err := s.GetOrganization(ctx, &GetOrgOptions{Name: opt.Organization}); err != nil {
+	if _, err := s.GetOrganization(ctx, &OrganizationOptions{Name: opt.Organization}); err != nil {
 		return nil, errors.New("organization not found")
 	}
 	if _, err := s.getRepository(&RepositoryOptions{
@@ -186,7 +186,7 @@ func (s *MockSCM) UpdateIssue(ctx context.Context, opt *IssueOptions) (*Issue, e
 	if !opt.valid() {
 		return nil, fmt.Errorf("invalid argument: %v", opt)
 	}
-	if _, err := s.GetOrganization(ctx, &GetOrgOptions{Name: opt.Organization}); err != nil {
+	if _, err := s.GetOrganization(ctx, &OrganizationOptions{Name: opt.Organization}); err != nil {
 		return nil, errors.New("organization not found")
 	}
 	if _, err := s.getRepository(&RepositoryOptions{
@@ -213,7 +213,7 @@ func (s *MockSCM) GetIssue(ctx context.Context, opt *RepositoryOptions, issueNum
 	if !opt.valid() {
 		return nil, fmt.Errorf("invalid argument: %v", opt)
 	}
-	if _, err := s.GetOrganization(ctx, &GetOrgOptions{Name: opt.Owner}); err != nil {
+	if _, err := s.GetOrganization(ctx, &OrganizationOptions{Name: opt.Owner}); err != nil {
 		return nil, errors.New("organization not found")
 	}
 	if _, err := s.getRepository(&RepositoryOptions{
@@ -234,7 +234,7 @@ func (s *MockSCM) GetIssues(ctx context.Context, opt *RepositoryOptions) ([]*Iss
 	if !opt.valid() {
 		return nil, fmt.Errorf("invalid argument: %v", opt)
 	}
-	if _, err := s.GetOrganization(ctx, &GetOrgOptions{Name: opt.Owner}); err != nil {
+	if _, err := s.GetOrganization(ctx, &OrganizationOptions{Name: opt.Owner}); err != nil {
 		return nil, errors.New("organization not found")
 	}
 	if _, err := s.getRepository(&RepositoryOptions{
@@ -257,7 +257,7 @@ func (s *MockSCM) DeleteIssue(ctx context.Context, opt *RepositoryOptions, issue
 	if !opt.valid() {
 		return fmt.Errorf("invalid argument: %v", opt)
 	}
-	if _, err := s.GetOrganization(ctx, &GetOrgOptions{Name: opt.Owner}); err != nil {
+	if _, err := s.GetOrganization(ctx, &OrganizationOptions{Name: opt.Owner}); err != nil {
 		return errors.New("organization not found")
 	}
 	if _, err := s.getRepository(&RepositoryOptions{
@@ -274,7 +274,7 @@ func (s *MockSCM) DeleteIssues(ctx context.Context, opt *RepositoryOptions) erro
 	if !opt.valid() {
 		return fmt.Errorf("invalid argument: %v", opt)
 	}
-	if _, err := s.GetOrganization(ctx, &GetOrgOptions{Name: opt.Owner}); err != nil {
+	if _, err := s.GetOrganization(ctx, &OrganizationOptions{Name: opt.Owner}); err != nil {
 		return errors.New("organization not found")
 	}
 	if _, err := s.getRepository(&RepositoryOptions{
@@ -296,7 +296,7 @@ func (s *MockSCM) CreateIssueComment(ctx context.Context, opt *IssueCommentOptio
 	if !opt.valid() {
 		return 0, fmt.Errorf("invalid argument: %v", opt)
 	}
-	if _, err := s.GetOrganization(ctx, &GetOrgOptions{Name: opt.Organization}); err != nil {
+	if _, err := s.GetOrganization(ctx, &OrganizationOptions{Name: opt.Organization}); err != nil {
 		return 0, errors.New("organization not found")
 	}
 	if _, err := s.getRepository(&RepositoryOptions{
@@ -318,7 +318,7 @@ func (s *MockSCM) UpdateIssueComment(ctx context.Context, opt *IssueCommentOptio
 	if !opt.valid() {
 		return fmt.Errorf("invalid argument: %v", opt)
 	}
-	if _, err := s.GetOrganization(ctx, &GetOrgOptions{Name: opt.Organization}); err != nil {
+	if _, err := s.GetOrganization(ctx, &OrganizationOptions{Name: opt.Organization}); err != nil {
 		return errors.New("organization not found")
 	}
 	if _, err := s.getRepository(&RepositoryOptions{
@@ -358,7 +358,7 @@ func (s *MockSCM) CreateCourse(ctx context.Context, opt *CourseOptions) ([]*Repo
 	if !opt.valid() {
 		return nil, fmt.Errorf("invalid argument: %v", opt)
 	}
-	org, err := s.GetOrganization(ctx, &GetOrgOptions{ID: opt.OrganizationID, NewCourse: true})
+	org, err := s.GetOrganization(ctx, &OrganizationOptions{ID: opt.OrganizationID, NewCourse: true})
 	if err != nil {
 		return nil, err
 	}
@@ -407,7 +407,7 @@ func (s *MockSCM) UpdateEnrollment(ctx context.Context, opt *UpdateEnrollmentOpt
 	if !opt.valid() {
 		return nil, fmt.Errorf("invalid argument: %v", opt)
 	}
-	org, err := s.GetOrganization(ctx, &GetOrgOptions{
+	org, err := s.GetOrganization(ctx, &OrganizationOptions{
 		Name: opt.Organization,
 	})
 	if err != nil {
@@ -418,7 +418,6 @@ func (s *MockSCM) UpdateEnrollment(ctx context.Context, opt *UpdateEnrollmentOpt
 			Organization: org.Name,
 			Path:         qf.StudentRepoName(opt.User),
 			Private:      true,
-			Owner:        org.Name,
 		})
 	}
 	return nil, nil
@@ -429,7 +428,7 @@ func (s *MockSCM) RejectEnrollment(ctx context.Context, opt *RejectEnrollmentOpt
 	if !opt.valid() {
 		return fmt.Errorf("invalid argument: %v", opt)
 	}
-	if _, err := s.GetOrganization(ctx, &GetOrgOptions{
+	if _, err := s.GetOrganization(ctx, &OrganizationOptions{
 		ID: opt.OrganizationID,
 	}); err != nil {
 		return errors.New("organization not found")
@@ -448,7 +447,7 @@ func (s *MockSCM) CreateGroup(ctx context.Context, opt *TeamOptions) (*Repositor
 	if !opt.valid() {
 		return nil, nil, fmt.Errorf("invalid argument: %v", opt)
 	}
-	if _, err := s.GetOrganization(ctx, &GetOrgOptions{
+	if _, err := s.GetOrganization(ctx, &OrganizationOptions{
 		Name: opt.Organization,
 	}); err != nil {
 		return nil, nil, errors.New("organization not found")
@@ -473,7 +472,7 @@ func (s *MockSCM) DeleteGroup(ctx context.Context, opt *GroupOptions) error {
 	if !opt.valid() {
 		return fmt.Errorf("invalid argument: %v", opt)
 	}
-	if _, err := s.GetOrganization(ctx, &GetOrgOptions{
+	if _, err := s.GetOrganization(ctx, &OrganizationOptions{
 		ID: opt.OrganizationID,
 	}); err != nil {
 		return errors.New("organization not found")
